write on server side

+++ NFSClient.h
@@ -40,8 +40,8 @@ public:
   int NFSPROC_MKNOD(const char *, mode_t, dev_t);
   int NFSPROC_OPEN(const char *, const struct fuse_file_info *, int *);
   int NFSPROC_RELEASE(const char *, const struct fuse_file_info *);
-  int NFSPROC_READ(const char *, char *, size_t, off_t, const struct fuse_file_info *, int *);
-  int NFSPROC_WRITE(const char *, const char *, size_t, off_t, const struct fuse_file_info *, int *);
+  int NFSPROC_READ(const char *, char *, size_t, off_t, const struct fuse_file_info *, ssize_t *);
+  int NFSPROC_WRITE(const char *, const char *, size_t, off_t, const struct fuse_file_info *, ssize_t *);
   int NFSPROC_FGETATTR(const char *, struct stat *, const struct fuse_file_info *);
   int NFSPROC_READDIR(const char *, void *, fuse_fill_dir_t, off_t, struct fuse_file_info *);
